CM - 1 Created function prototypes for weekly report verification Created two function prototypes get_volunteer_info and verify_weekly_report to help implementing the features to verfify the validity of a given weekly report.
The firstion function will retrieve volunteer data from the main tracking
google sheet, and calculate all necessary information for further usage.

The second function will be used to verify if a given weekly report is
valid, based on if there exist a the record in the report for the current
week.